requirements.txt
colored==1.4.2
numpy==1.18.0
Pillow==6.2.1
psutil==5.6.7
PyYAML==5.2
six==1.13.0
wcwidth==0.1.7
wxPython==4.0.7.post2
